Updated the layout for creating a prognosis
There is now a layout to support viewing the norms that have effect on this prognosis. There is also a edit button that can be implemented to edit weather impacts. Lastly excess tab now has a button to open the comming excess view that allows you to view the text in a bigger area and set the impact all in one view.