PR request fix (ish)
Was unable to get Edit view to show data correctly (without the trailing zeroes)

Still have to move logic to helper class, but that is not priority right now.

Input fields now accept a single decimal, or specifically -.25 or -.75

Also checks that minutes do not allow decimals. Was already correct for going to database, now user also knows